djangoapi/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
